Update tex_offset to int64 (#40)

+++ src/bridge/mgr.cpp
@@ -475,7 +475,7 @@ static RTAssets loadRenderObjects(
     SourceTexture *out_textures = tmp_alloc.allocN<SourceTexture>(model.numTextures);
 
     for (CountT i = 0; i < model.numTextures; i++) {
-        uint32_t tex_offset = model.texOffsets[i];
+        uint64_t tex_offset = model.texOffsets[i];
         Optional<SourceTexture> tex = SourceTexture {
             .data = &model.texData[tex_offset],
             .format = SourceTextureFormat::R8G8B8A8,
